+++ h5netcdf/tests/test_h5netcdf.py
@@ -1467,3 +1467,39 @@ def test_group_names(tmp_local_netcdf):
         for i in range(10):
             name = "/".join([name, f"group{i:02d}"])
             assert ds[name].name == name
+
+
+def test_legacyapi_endianess(tmp_local_netcdf):
+    big = legacyapi._check_return_dtype_endianess("big")
+    little = legacyapi._check_return_dtype_endianess("little")
+    native = legacyapi._check_return_dtype_endianess("native")
+
+    with legacyapi.Dataset(tmp_local_netcdf, "w") as ds:
+        ds.createDimension("x", 4)
+        # test creating variable using endian keyword argument
+        v = ds.createVariable("big", int, ("x"), endian="big")
+        v[...] = 65533
+        v = ds.createVariable("little", int, ("x"), endian="little")
+        v[...] = 65533
+        v = ds.createVariable("native", int, ("x"), endian="native")
+        v[...] = 65535
+
+    with h5py.File(tmp_local_netcdf, "r") as ds:
+        assert ds["big"].dtype.byteorder == big
+        assert ds["little"].dtype.byteorder == little
+        assert ds["native"].dtype.byteorder == native
+
+    with h5netcdf.File(tmp_local_netcdf, "r") as ds:
+        assert ds["big"].dtype.byteorder == big
+        assert ds["little"].dtype.byteorder == little
+        assert ds["native"].dtype.byteorder == native
+
+    with legacyapi.Dataset(tmp_local_netcdf, "r") as ds:
+        assert ds["big"].dtype.byteorder == big
+        assert ds["little"].dtype.byteorder == little
+        assert ds["native"].dtype.byteorder == native
+
+    with netCDF4.Dataset(tmp_local_netcdf, "r") as ds:
+        assert ds["big"].dtype.byteorder == big
+        assert ds["little"].dtype.byteorder == little
+        assert ds["native"].dtype.byteorder == native